JAX-RS implementation using Apache CXF
