+++ computer-database-webservice-sdk/.gitignore
@@ -1,0 +1,1 @@
+/target/
