+++ computer-database-webservice-sdk/pom.xml
@@ -1,0 +1,25 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+	<parent>
+		<groupId>computer-database</groupId>
+		<artifactId>computer-database</artifactId>
+		<version>0.3.1</version>
+	</parent>
+	<artifactId>computer-database-webservice-sdk</artifactId>
+	<dependencies>
+		<dependency>
+			<groupId>javax.ws.rs</groupId>
+			<artifactId>javax.ws.rs-api</artifactId>
+			<version>${javax-rs-api.version}</version>
+			<scope>provided</scope>
+		</dependency>
+		
+		<!-- Internal Dependencies -->
+		<dependency>
+			<groupId>${project.groupId}</groupId>
+			<artifactId>computer-database-binding</artifactId>
+			<version>${project.version}</version>
+		</dependency>
+	</dependencies>
+</project>
