+++ computer-database-webservice/src/main/webapp/WEB-INF/web.xml
@@ -12,11 +12,7 @@
   
   <servlet>
     <servlet-name>cdbWebService</servlet-name>
-    <servlet-class>org.glassfish.jersey.servlet.ServletContainer</servlet-class>
-    <init-param>
-        <param-name>jersey.config.server.provider.packages</param-name>
-        <param-value>fr.heffebaycay.cdb.webservice</param-value>
-    </init-param>
+    <servlet-class>org.apache.cxf.transport.servlet.CXFServlet</servlet-class>
     <load-on-startup>1</load-on-startup>
   </servlet>
   <servlet-mapping>
