+++ pom.xml
@@ -12,8 +12,9 @@
     <module>computer-database-service</module>
     <module>computer-database-binding</module>
     <module>computer-database-webapp</module>
-    <module>computer-database-console</module>
+	<module>computer-database-webservice-sdk</module>
     <module>computer-database-webservice</module>
+	<module>computer-database-console</module>
   </modules>
 
   <properties>
@@ -33,9 +34,10 @@
   	<spring.version>4.1.4.RELEASE</spring.version>
   	<validation-api.version>1.1.0.Final</validation-api.version>
   	<hibernate-validator.version>5.1.3.Final</hibernate-validator.version>
-	<jersey.version>2.16</jersey.version>
+	<javax-rs-api.version>2.0.1</javax-rs-api.version>
+	<apache-cxf-jaxrs.version>3.0.3</apache-cxf-jaxrs.version>
 	<jackson-jaxrs-json-provider.version>2.5.1</jackson-jaxrs-json-provider.version>
-        <spring-security.version>3.2.5.RELEASE</spring-security.version>
+    <spring-security.version>3.2.5.RELEASE</spring-security.version>
     <apache-commons-lang.version>3.3.2</apache-commons-lang.version>
     <apache-commons-validator>1.4.1</apache-commons-validator>
     <hibernate-orm.version>4.3.8.Final</hibernate-orm.version>
